move Library and Strategies into EDIT_ME, move bps_without_dupes from strategies to library
