+++ EDIT_ME/Strategies.ts
@@ -1,84 +1,4 @@
-import { ProductName } from '../types/Blueprint';
-import { SetMergingStrategy } from '../types/MergingStrategy';
-
-// TODO: GH-4: only merge if there are excess blueprints for that product. for now, hardcode ones the algo wants that i don't have
-export const BPS_WITHOUT_DUPES: ProductName[] = [
-  // // income wants
-  'Bongos',
-  'Harp',
-  'Tambourine', // in MWS but not in a regular pack
-  'Viola',
-  'Harpsichord',
-  'Snare Drum',
-  'Saxophone',
-  // 'Microscope',
-  // 'Machine Parts',
-  // 'Motor Unit',
-  // 'Telephone',
-  // 'Steam Boat',
-  // 'Locomotive',
-  // 'Mechanical Parts', // not event only
-  // 'Electro Magnet',
-  // 'Clockwork',
-  // 'Steam Engine',
-  // 'Camera',
-  // 'Clock',
-  // 'Electrical Parts', // not event only. technically.
-  'Combustion Engine',
-  'Glider',
-  'Bicycle',
-  'Truck',
-  'Motorcycle',
-  'Car',
-  'Airplane',
-  'Antenna',
-  'Movie Projector',
-  'Walkie Talkie',
-  'TV Set',
-  'Microchip',
-  'Radio',
-  'Monitor',
-  'PC',
-
-  // research
-  // modern technology
-  'Stethoscope',
-  'Lamp',
-  'Refrigerator',
-  'X-ray Machine',
-  'Flashlight',
-  'Kettle Grill',
-  'Stove',
-  'Defibrillator',
-  'Light Bulb', // in MWS but not in a regular pack
-
-  // science tools - tools pack & more
-  // 'Mechanical Parts',
-  // 'Compass',
-  // 'Sickle',
-  // 'Chisel',
-
-  // modern exploration
-  'Antenna',
-  'Diving Gear',
-  'Rocket',
-  'Tent',
-  'Submarine',
-  'Space Probe',
-
-  // exploration - in master & grandmaster packs
-  // 'Paper',
-  // 'Paper Sheets',
-
-  // // gems
-  // 'Cut Emerald',
-  // 'Cut Ruby',
-  // 'Cut Sapphire',
-  // 'Cut Onyx',
-
-  // 'Lump Hammer',
-  // 'War Hammer',
-];
+import { SetMergingStrategy } from '../src/upgradeBlueprints/types/MergingStrategy';
 
 export const STRATEGIES: SetMergingStrategy[] = [
   {
